Add support for provisioning and configuring a bastion host
...based on the scripts provided in the HSL server-based repo:
https://gitlab.hsl.fi/platforms/server-based

Developers can log into the host using a public key, which has been
signed using the shared developer private key (stored in the common
Azure key vault).

Instructions for signing your own key as a developer in order to log
into the bastion host can be found here:
https://gitlab.hsl.fi/developer-resources/azure-ansible/-/tree/master#signing-users-public-key-with-private-ca-key

The CA key is stored in the hsl-jore4-vault, which is common to all
environments.