Protect prod and test environments from accidental deletion
Apply a "lock: cannotdelete" -tag to the prod and test resource groups
preventing accidental deletion of resources.

In order to delete resources, use the Owner-role to remove the lock. If
the tag is not removed too, the lock will be re-applied within 24 hours.

For more details on how to use the Owner-role and on the lock-tag, see
the HSL Azure dashboard:
https://portal.azure.com/#@hslfi.onmicrosoft.com/dashboard/arm/subscriptions/b13714ed-2c1b-416c-89a9-909524515193/resourcegroups/dashboards/providers/microsoft.portal/dashboards/bcea8162-492c-4428-ba8c-19321eceb0cd