Apply routing table to all subnets
This is required in order to be able to control the traffic between the
jore4 vnet and the peered vnet using a firewall (in the peered vnet).

If no routing table is applied in a certain vnet, then all traffic in
that subnet is freely routed to the peered vnet.

This has to be taken into account when adding new subnets. The jore4
routing table has to be applied also to all new subnets.